Add compression
With this step the binary size drops from 78 MB to 30 MB. Unfortunately,
at the cost of startup time -- but since the previous commit just sped
it up by 50-60%, we are back to where we were.

I don't think the initial startup time being in the range of 100-300 ms
is really all that big of a deal; for an interactive session the
following lookups are lightning fast.

Fixes #21.