Re-derive Serialize for entries
I don't think anyone else is using Jisho as a library yet, so no harm
done, but removing this as part of switching from JSON to Bitcode was of
course a breaking change.

I've now decided that I *do* want to support easy serialization via
Serde, so from now on it's a supported feature.